feat(organization): 实现字典数据实体独立字段并修复行尾换行符
- 将 DictDataEntity 从继承 BaseEntity 改为直接实现 Serializable，
  独立声明 id、createTime、updateTime 等字段，以适配 sys_dict_data 表结构
- 统一多个 Java 文件的换行符为 LF，消除 CRLF 差异
- 调整 DictController、DictServiceImpl、DeptServiceImpl、PostController 等文件格式

+++ backend/hrms-system/hrms-system-organization/src/main/java/com/hrms/system/organization/service/impl/PostServiceImpl.java
@@ -62,7 +62,7 @@ public class PostServiceImpl implements PostService {
                 .map(this::convertToVO)
                 .collect(Collectors.toList());
 
-        return PageResult.of(records, resultPage.getTotal(), resultPage.getCurrent(), resultPage.getSize());
+        return PageResult.of(records, resultPage.getTotal(), (int) resultPage.getCurrent(), (int) resultPage.getSize());
     }
 
     @Override
